third_party/blink/web_tests/platform/generic/fast/dom/MutationObserver/takeRecords-expected.txt

Testing WebKitMutationObserver.takeRecords

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Testing takeRecords.
...records are taken synchronously.
PASS mutations.length is 2
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "foo"
PASS mutations[0].attributeNamespace is null
PASS mutations[1].type is "characterData"
PASS mutations[1].target is subDiv.firstChild
...takeRecord took records, but did not clear transient observers
PASS mutations.length is 1
PASS mutations[0].type is "attributes"
PASS mutations[0].target is subDiv
PASS mutations[0].attributeName is "foo"


PASS successfullyParsed is true

TEST COMPLETE
